Hermes bigger workloads + automatic cluster runs

+++ include/mica/mica.h
@@ -96,8 +96,10 @@ struct mica_kv {
 	int instance_id;	/* ID of this MICA instance. Used for shm keys */
 	int node_id;
 
-	int num_bkts;	/* Number of buckets requested by user */
-	int bkt_mask;	/* Mask down from a mica_key's @bkt to a bucket */
+    uint64_t num_bkts;	/* Number of buckets requested by user */
+    uint64_t bkt_mask;	/* Mask down from a mica_key's @bkt to a bucket */
+//	int num_bkts;	/* Number of buckets requested by user */
+//	int bkt_mask;	/* Mask down from a mica_key's @bkt to a bucket */
 
 	uint64_t log_cap;	/* Capacity of circular log in bytes */
 	uint64_t log_mask;	/* Mask down from a slot's @offset to a log offset */
@@ -127,7 +129,7 @@ struct wrkr_ud_req {
 };
 
 void mica_init(struct mica_kv *kv,
-	int instance_id, int node_id, int num_bkts, u_int64_t log_cap);
+	int instance_id, int node_id, int num_bkts, uint64_t log_cap);
 
 /* Single-key INSERT */
 void mica_insert_one(struct mica_kv *kv,
